requirements.txt
streamlit>=1.32.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
pypdf>=4.0.0
python-docx>=1.1.0
pdfplumber>=0.10.0
plotly>=5.18.0
google-generativeai>=0.5.0
lxml>=5.1.0